% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{addBalance}
\alias{addBalance}
\title{Add fine balance edges}
\usage{
addBalance(net, treatedVals, controlVals, replaceExisting = TRUE)
}
\arguments{
\item{net}{the network object created for the network flow problem}

\item{treatedVals}{the balance value for treated nodes}

\item{controlVals}{the balance value for control nodes}

\item{replaceExisting}{(optional) whether or not to replace the existing net;
TRUE by default}
}
\value{
the network structure with balance edges added
}
\description{
Add fine balance edges
}
