% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{addExclusion}
\alias{addExclusion}
\title{Add exclusion edges}
\usage{
addExclusion(net, remove = FALSE)
}
\arguments{
\item{net}{the input network structure}

\item{remove}{(optional) whether to exclude edges; FALSE by default}
}
\value{
the network structure with exclusion edges added to allow for trdeoff
for the exclusion cost
}
\description{
Add exclusion edges
}
