% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{balanceCosts}
\alias{balanceCosts}
\title{Create a skeleton representation of the balance edge costs}
\usage{
balanceCosts(net, balance.penalty = 1)
}
\arguments{
\item{net}{the network structure}

\item{balance.penalty}{(optional) the numeric value for balance; 1 by default}
}
\value{
the skeleton with balance edge cost
}
\description{
Create a skeleton representation of the balance edge costs
associated with pairings for a given distance and network
}
