% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_funcs.R
\name{build.dist.struct_user}
\alias{build.dist.struct_user}
\title{An internal helper function that generates the data abstraction for the edge
weights of the main network structure using the distance matrix passed by the
user.}
\usage{
build.dist.struct_user(z, distMat, verbose = FALSE)
}
\arguments{
\item{z}{a vector indicating whether each unit is in treatment or control
group}

\item{distMat}{a matrix of pair-wise distance}

\item{verbose}{a boolean value whether to print(cat) debug information.
Default: FALSE}
}
\value{
a distance structure used for constructing the main network flow
problem
}
\description{
An internal helper function that generates the data abstraction for the edge
weights of the main network structure using the distance matrix passed by the
user.
}
