% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_funcs.R
\name{combine_dist}
\alias{combine_dist}
\title{An internal helper function that combines two distance object}
\usage{
combine_dist(a, b)
}
\arguments{
\item{a}{a distance structure object}

\item{b}{a distance structure object}
}
\value{
a new distance structure object whose edge weights are the sum of the
corresponding edge weigths in a and b
}
\description{
An internal helper function that combines two distance object
}
