% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{combine_match_result}
\alias{combine_match_result}
\title{Combine two matching result}
\usage{
combine_match_result(matching_result1, matching_result2)
}
\arguments{
\item{matching_result1}{the first matching result object.}

\item{matching_result2}{the second matching result object.}
}
\value{
a new matching result combining two objects. Note that
the matching index for the second matching is the original name
plus the maximum match index in the first matching object.
}
\description{
Combine two matching result
}
