% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{convert_index}
\alias{convert_index}
\title{An internal helper function that translates the matching index in the sorted
data frame to the original dataframe's row index}
\usage{
convert_index(matching_result)
}
\arguments{
\item{matching_result}{an object returned by the main matching function
dist_bal_match}
}
\description{
An internal helper function that translates the matching index in the sorted
data frame to the original dataframe's row index
}
