% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_funcs.R
\name{data_precheck}
\alias{data_precheck}
\title{Data precheck: Handle missing data(mean imputation) and remove redundant
columns; it also adds an NA column for indicating whether it's missing}
\usage{
data_precheck(X)
}
\arguments{
\item{X}{a dataframe that the user initially inputs for matching - dataframe
with covariates}
}
\value{
a dataframe with modified data if necessary
}
\description{
Data precheck: Handle missing data(mean imputation) and remove redundant
columns; it also adds an NA column for indicating whether it's missing
}
