% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr_stats.R
\name{descr.stats_general}
\alias{descr.stats_general}
\title{Generate summary statistics for matches}
\usage{
descr.stats_general(matches, df, treatCol, b.vars, pair.vars, extra = FALSE)
}
\arguments{
\item{matches}{One matching result from the main matching function}

\item{df}{the original data frame used for matching}

\item{treatCol}{the character of the column name for treatment vector}

\item{b.vars}{the vector of column names of covariates used for measuring
balance}

\item{pair.vars}{the vector of column names used for measuring pairwise
distance}

\item{extra}{the list of summary statistic; it must be the types that can be
taken by cobalt}
}
\value{
a named vector of summary statistic
}
\description{
Generate summary statistics for matches
}
