% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{distanceFunctionHelper}
\alias{distanceFunctionHelper}
\title{Helper function that change input distance matrix}
\usage{
distanceFunctionHelper(z, distMat)
}
\arguments{
\item{z}{the treatment vector}

\item{distMat}{the user input distance matrix}
}
\value{
a distance matrix where (i,j) element is the distance between unit i
and j in the same order as z
}
\description{
Helper function that change input distance matrix
}
