% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_dummy.R
\name{dummy}
\alias{dummy}
\title{This is a modified version of the function "dummy" from the R package
dummies. Original code Copyright (c) 2011 Decision
Patterns.}
\usage{
dummy(
  x,
  data = NULL,
  sep = "",
  drop = TRUE,
  fun = as.integer,
  verbose = FALSE,
  name = NULL
)
}
\arguments{
\item{x}{a data.frame, matrix or single variable or variable name}

\item{data}{(optional) if provided, x is the name of a column on the data}

\item{sep}{(optional) the separator used between variable name and the value}

\item{drop}{(optional) whether to drop unused levels}

\item{fun}{(optional) function to coerce the value in the final matrix; 'as,integer' by default}

\item{verbose}{(optional) whether to print the number of variables; FALSE by default}

\item{name}{(optional) the column name to be selected for converting; NULL by default}
}
\description{
Change is made to the "model.matrix" function so that the output could be
used for the current package.
}
