% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{edgelist2ISM}
\alias{edgelist2ISM}
\title{Change the edgelist to the infinity sparse matrix}
\usage{
edgelist2ISM(elist)
}
\arguments{
\item{elist}{the vector of the edges}
}
\value{
the infinity sparse matrix object
}
\description{
Change the edgelist to the infinity sparse matrix
}
