% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{excludeCosts}
\alias{excludeCosts}
\title{Create a skeleton representation of the exclusion edge costs}
\usage{
excludeCosts(net, exclude.penalty = 1)
}
\arguments{
\item{net}{the network structure}

\item{exclude.penalty}{(optional) numeric penalty for excluding a treated
unit; 1 by default}
}
\value{
the skeleton with exclusion edge cost
}
\description{
Create a skeleton representation of the exclusion edge costs
associated with pairings for a given distance and network
}
