% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{extractEdges}
\alias{extractEdges}
\title{Extract edges from the network}
\usage{
extractEdges(net)
}
\arguments{
\item{net}{the network representation}
}
\value{
the list of edges
}
\description{
Extract edges from the network
}
