% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{extractSupply}
\alias{extractSupply}
\title{Extract the supply nodes from the net}
\usage{
extractSupply(net)
}
\arguments{
\item{net}{the network representation}
}
\value{
the vector of the supply nodes
}
\description{
Extract the supply nodes from the net
}
