% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{filter_match_result}
\alias{filter_match_result}
\title{Filter match result}
\usage{
filter_match_result(matching_result, filter_expr)
}
\arguments{
\item{matching_result}{the matching result object.}

\item{filter_expr}{character, the filtering condition based on the
summary table returned by get_rho_obj.}
}
\value{
the filtered match result
}
\description{
Filter match result
}
