% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{flattenSkeleton}
\alias{flattenSkeleton}
\title{Turns a skeleton representation of edge costs in a network}
\usage{
flattenSkeleton(skeleton)
}
\arguments{
\item{skeleton}{the skeleton structure}
}
\value{
vector representation expected by the optimization routine.
}
\description{
Turns a skeleton representation of edge costs in a network back
into the vector representation expected by the optimization routine. See
comment on the costSkeleton function for more details.
}
