% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{generate_absdiff_tv_graph}
\alias{generate_absdiff_tv_graph}
\title{Plotting function that generate absolute standardized difference of balance variable v.s. sum of pair-wise distance}
\usage{
generate_absdiff_tv_graph(matchingResult)
}
\arguments{
\item{matchingResult}{a matching result object containing information for all matches}
}
\description{
Plotting function that generate absolute standardized difference of balance variable v.s. sum of pair-wise distance
}
