% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{generate_rhos}
\alias{generate_rhos}
\title{Generate rho pairs}
\usage{
generate_rhos(rho1.list, rho2.list)
}
\arguments{
\item{rho1.list}{a vector of rho 1 values}

\item{rho2.list}{a vector of rho 2 values}
}
\value{
a vector of (rho1, rho2) pairs
}
\description{
An internal helper function that generates the set of rho value
pairs used for matching. This function is used when exploring the Pareto
optimality of the solutions to the multi-objective optimization in matching.
}
