% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{generate_tv_graph}
\alias{generate_tv_graph}
\title{Plotting function that generate the total variation imbalance v.s. number of unmatched treated units}
\usage{
generate_tv_graph(matchingResult)
}
\arguments{
\item{matchingResult}{a matching result object containing information for all matches}
}
\description{
Plotting function that generate the total variation imbalance v.s. number of unmatched treated units
}
