% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{get_five_index}
\alias{get_five_index}
\title{An internal helper function that gives the index of matching with a wide
range of number of treated units left unmatched}
\usage{
get_five_index(matching_result)
}
\arguments{
\item{matching_result}{an object returned by the main matching function
dist_bal_match}
}
\value{
a vector of five matching indices with the number of treated units
excluded at 0th, 25th, 50th, 75th and 100th percentiles respectively.
}
\description{
An internal helper function that gives the index of matching with a wide
range of number of treated units left unmatched
}
