% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{get_rho_obj}
\alias{get_rho_obj}
\title{Penalty and objective values summary}
\usage{
get_rho_obj(matching_result)
}
\arguments{
\item{matching_result}{matchingResult object that contains information for all
matches.}
}
\value{
a dataframe that contains objective function values and rho values
corresponding coefficients before each objective function.
}
\description{
Helper function to generate a dataframe with matching number,
penalty (rho) values, and objective function values.
}
\seealso{
Other numerical analysis helper functions: 
\code{\link{generateRhoObj}()},
\code{\link{get_balance_table}()},
\code{\link{get_unmatched}()}
}
\concept{numerical analysis helper functions}
