% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{get_unmatched}
\alias{get_unmatched}
\title{Get unmatched percentage}
\usage{
get_unmatched(matching_result)
}
\arguments{
\item{matching_result}{matchingResult object that contains information for all
matches}
}
\value{
data frame with three columns, one containing the matching index, one
containing the number of matched units, and one conatining the percentage
of matched units (out of original treated group size).
}
\description{
A function that generate the percentage of unmatched units for
each match.
}
\examples{
\dontrun{
get_unmatched(match_result)
}
}
\seealso{
Other numerical analysis helper functions: 
\code{\link{generateRhoObj}()},
\code{\link{get_balance_table}()},
\code{\link{get_rho_obj}()}
}
\concept{numerical analysis helper functions}
