% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{getExactOn}
\alias{getExactOn}
\title{Generate a factor for exact matching.}
\usage{
getExactOn(dat, exactList)
}
\arguments{
\item{dat}{dataframe containing all the variables in exactList}

\item{exactList}{factor of names of the variables on which we want exact or
close matching.}
}
\value{
factor on which to match exactly, with labels given by concatenating
labels for input variables.
}
\description{
Generate a factor for exact matching.
}
