% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{getPropensityScore}
\alias{getPropensityScore}
\title{Fit propensity scores using logistic regression.}
\usage{
getPropensityScore(data, covs)
}
\arguments{
\item{data}{dataframe that contains a column named "treat", the treatment
vector, and columns of covariates specified.}

\item{covs}{factor of column names of covariates used for fitting a
propensity score model.}
}
\value{
vector of estimated propensity scores (on the probability scale).
}
\description{
Fit propensity scores using logistic regression.
}
