% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{makeInfinitySparseMatrix}
\alias{makeInfinitySparseMatrix}
\title{Internal helper to build infinity sparse matrix}
\usage{
makeInfinitySparseMatrix(
  data,
  cols,
  rows,
  colnames = NULL,
  rownames = NULL,
  dimension = NULL,
  call = NULL
)
}
\arguments{
\item{data}{the input numeric vector of cost}

\item{cols}{the input numeric vector corresponding to control units}

\item{rows}{the input numeric vector corresponding to treated units}

\item{colnames}{(optional) vector containing names for all control units}

\item{rownames}{(optional) vectpr containing names for all treated units}

\item{dimension}{(optional) vector of number of treated and control units}

\item{call}{(optional) funtion call used to create the InfinitySparseMatrix}
}
\value{
an InfinitySparseMatrix object
}
\description{
Formats the data and make a call to
\link[optmatch]{InfinitySparseMatrix-class}
}
