% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{makeSparse}
\alias{makeSparse}
\title{Helper function to mask edges}
\usage{
makeSparse(net, mask, replaceMask = TRUE)
}
\arguments{
\item{net}{the network object}

\item{mask}{a matrix indicating whether to exclude the corresponding edge}

\item{replaceMask}{(optional) whether to mask}
}
\value{
the masked network structure object
}
\description{
Remove some of the treatment-control edges from a network flow
representation of a match (forbidding those pairings)
}
