% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{matched_index}
\alias{matched_index}
\title{An internal helper function that translate the matching index in the sorted
data frame to the original dataframe's row index}
\usage{
matched_index(matchingResult)
}
\arguments{
\item{matchingResult}{an object returned by the main matching function
dist_bal_match}
}
\description{
An internal helper function that translate the matching index in the sorted
data frame to the original dataframe's row index
}
