% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{matrix2cost}
\alias{matrix2cost}
\title{change the distance matrix to cost}
\usage{
matrix2cost(net, distance)
}
\arguments{
\item{net}{the network structure}

\item{distance}{distance matrix}
}
\value{
the vector of cost
}
\description{
change the distance matrix to cost
}
