% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{meldMask}
\alias{meldMask}
\title{Helper function to combine two sparse distances}
\usage{
meldMask(mask1, mask2)
}
\arguments{
\item{mask1}{matrix of the first mask}

\item{mask2}{matrix of the second mask}
}
\value{
combined mask structure
}
\description{
Combine two sparse distances,
allowing only pairings allowed by both
}
