% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{multiObjMatch}
\alias{multiObjMatch}
\title{The main matching function that can return a matching object}
\usage{
multiObjMatch(
  df,
  treatCol,
  responseCol,
  distList,
  exactlist,
  myBalCol,
  rho1 = c(1),
  rho2 = c(1, 2, 3),
  propensityCols = NULL,
  pScores = NULL,
  idCol = NULL,
  maxUnMatched = 0.25,
  caliperOption = 0.25,
  toleranceOption = 0.01,
  maxIter = 0,
  rho.max.f = 10
)
}
\arguments{
\item{df}{dataframe that contain columns indicating treatment, outcome and covariates}

\item{treatCol}{character of name of the column indicating treatment assignment}

\item{responseCol}{character of name of the column indicating the outcome variable}

\item{distList}{factor of the names of the}

\item{exactlist}{factor of the names of the variables that we want exact matching on}

\item{myBalCol}{charactor of column name of the variable that we want to evaluate balance on}

\item{rho1}{factor of values of rho1. Default value is c(1).}

\item{rho2}{factor of values of rho2. Default value is c(1,2,3)}

\item{propensityCols}{factor of names of columns that users want to fit a propensity score model}

\item{pScores}{character of the name of the column that indicate the propensity score; default is NULL}

\item{idCol}{character of the name of the column that indicate the id for each unit; default is NULL}

\item{maxUnMatched}{double of the maximum proportion of unmatched unit that can be accepted; default is 0.25}

\item{caliperOption}{double of the caliper value; default is 0.25}

\item{toleranceOption}{double of tolerance of close match distance; default is 1e-2}

\item{maxIter}{interger of the maximum number of iterations to search for (rho1, rho2) pair that improve the matching}

\item{rho.max.f}{double of the scaling factor used in proposal for rhos; default is 10}
}
\value{
a matching object
}
\description{
The main matching function that can return a matching object
}
