% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{netFlowMatch}
\alias{netFlowMatch}
\title{Create network flow structure}
\usage{
netFlowMatch(z, IDs = NULL)
}
\arguments{
\item{z}{a vector of treatment vectors}

\item{IDs}{(optional) the name of the units}
}
\value{
a networks structure
}
\description{
Create network flow structure
}
