% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{obj.to.match}
\alias{obj.to.match}
\title{An internal helper function that transforms the output from the RELAX
algorithm to a data structure that is more interpretable for the output of
the main matching function}
\usage{
obj.to.match(out.elem, already.done = NULL, prev.obj = NULL)
}
\arguments{
\item{out.elem}{a named list whose elements are: (1) the net structure (2)
the edge weights of pair-wise distance (3) the edge weights of marginal
balance (4) the list of rho value pairs (5) the named list of solutions
from the RELAX algorithm}

\item{already.done}{a factor indicating the index of matches already been
transformed}

\item{prev.obj}{an object of previously transformed matches}
}
\value{
a named list with elements containing matching information useful for
the main matching function
}
\description{
An internal helper function that transforms the output from the RELAX
algorithm to a data structure that is more interpretable for the output of
the main matching function
}
