% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{pairCosts}
\alias{pairCosts}
\title{Create a skeleton representation of the edge costs}
\usage{
pairCosts(dist.struct, net)
}
\arguments{
\item{dist.struct}{the distance structure}

\item{net}{the net structure}
}
\value{
the skeleton representation of the given distance pairs and the net
}
\description{
Create a skeleton representation of the edge costs
}
