% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{solveP}
\alias{solveP}
\title{Solve the network flow problem - basic version}
\usage{
solveP(net, f1.list, f2.list, rho, tol = 1e-05)
}
\arguments{
\item{net}{the network representation}

\item{f1.list}{the list of the first objective functions values for each node}

\item{f2.list}{the list of the second objective functions values for each
node}

\item{rho}{the penalty coefficient}

\item{tol}{the tolerance value for precision}
}
\value{
the solution represented in a named list
}
\description{
Solve the network flow problem - basic version
}
