% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_functions.R
\name{solveP1}
\alias{solveP1}
\title{Solve the network flow problem - twoDistMatch}
\usage{
solveP1(net, f1.list, f2.list, f3.list, rho1, rho2 = 0, tol = 1e-05)
}
\arguments{
\item{net}{the network representation}

\item{f1.list}{the list of the first objective functions values for
each node}

\item{f2.list}{the list of the second objective functions values for
each node}

\item{f3.list}{the list of the third objective functions values for
each node}

\item{rho1}{the penalty coefficient for the second objective}

\item{rho2}{the penalty coefficient for the third objective}

\item{tol}{the tolerance value for precision}
}
\value{
the solution represented in a named list
}
\description{
Solve the network flow problem - twoDistMatch
}
