% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{visualize}
\alias{visualize}
\title{Visualize tradeoffs}
\usage{
visualize(
  matchingResult,
  x_axis = "dist1",
  y_axis = "dist2",
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  display_all = FALSE,
  cond = NULL,
  xlim = NULL,
  ylim = NULL,
  display_index = TRUE,
  average_cost = FALSE
)
}
\arguments{
\item{matchingResult}{the matching result returned by either distBalMatch or
twoDistMatch.}

\item{x_axis}{character, naming the objective function shown on x-axis; one
of ("pair", "marginal", "dist1", "dist2", "exclude"), "dist1" by default.}

\item{y_axis}{character, naming the objective function shown on y-axis; one
of ("pair", "marginal", "dist1", "dist2", "exclude"), "dist2" by default.}

\item{xlab}{(optional) the axis label for x-axis; NULL by default.}

\item{ylab}{(optional) the axis label for y-axis; NULL by default.}

\item{main}{(optional) the title of the graph; NULL by default.}

\item{display_all}{(optional) whether to show all the labels for match index;
FALSE by default, which indicates the visualization function only labels
matches at quantiles of number of treated units being excluded.}

\item{cond}{(optional) NULL by default, which denotes all the matches are
shown; otherwise, takes a list of boolean values indicating whether to
include each match}

\item{xlim}{(optional) NULL by default; function automatically takes the max
of the first objective function values being plotted on x-axis;
if specified otherwise, pass in the numeric vector
c(lower_bound, upper_bound)}

\item{ylim}{(optional) NULL by default; function automatically takes the max
of the first objective function values being plotted on y-axis;
if specified otherwise, pass in the numeric vector
c(lower_bound, upper_bound)}

\item{display_index}{(optional) TRUE by default; whether to display match
index}

\item{average_cost}{(optional) FALSE by default; whether to show mean cost}
}
\value{
No return value, called for visualization of match result
}
\description{
Main visualization functions for showing the tradeoffs between
two of the three objective functions.
}
\details{
By default, the plotting function will show the tradeoff between
the first distance objective function and the marginal balance (if
distBalMatch) is used; or simply the second distance objective function, if
twoDistMatch is used.
}
\examples{
## Generate matches
data("lalonde", package="cobalt")
psCols <- c("age", "educ", "married", "nodegree")
treatVal <- "treat"
responseVal <- "re78"
pairDistVal <- c("age", "married", "educ", "nodegree")
exactVal <- c("educ")
myBalVal <- c("race")
r1s <- c( 0.1, 0.3, 0.5, 0.7, 0.9,1, 1.1, 1.2, 1.3, 1.4, 1.5, 1.6, 1.7)
r2s <- c(0.01)
matchResult <- distBalMatch(df=lalonde, treatCol=treatVal, myBalCol=myBalVal,
rhoExclude =r1s, rhoBalance=r2s,
distList=pairDistVal, exactlist=exactVal,
propensityCols = psCols,ignore = c(responseVal), maxUnMatched = 0.1,
caliperOption=NULL, toleranceOption=1e-1, maxIter=0, rho.max.f = 10)

## Visualization
visualize(matchResult, "marginal", "exclude")
visualize(matchResult, "pair", "exclude")
}
