% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{compare_matching}
\alias{compare_matching}
\title{Generate a table that compare the covariate balance across different possible matches}
\usage{
compare_matching(
  matchingResult,
  covList = NULL,
  display_option = FALSE,
  stat = "mean.diff"
)
}
\arguments{
\item{matchingResult}{an object returned by the main matching function multiObjMatch}

\item{covList}{factor of names of covariates that we want to evaluate covariate balance on; default is NULL. When set to Null, the program will compare the covariates that have been used to construct a propensity model}

\item{display_option}{boolean value of whether to display all the matches; default is FALSE, where matches at each quantile is displayed}

\item{stat}{character of the name of the statistic used for measuring covariate balance; default is "mean.diff". This argument is the same as used in "cobalt" package}
}
\value{
a dataframe that shows covariate balance in different matches
}
\description{
Generate a table that compare the covariate balance across different possible matches
}
