% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{convert_index}
\alias{convert_index}
\title{An internal helper function that translates the matching index in the sorted
data frame to the original dataframe's row index}
\usage{
convert_index(matchingResult)
}
\arguments{
\item{matchingResult}{an object returned by the main matching function
distBalMatch}
}
\description{
An internal helper function that translates the matching index in the sorted
data frame to the original dataframe's row index
}
