% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{convert_names}
\alias{convert_names}
\title{Internal helper function that converts axis name to internal variable name}
\usage{
convert_names(x, y)
}
\arguments{
\item{x}{the user input character for x-axis value}

\item{y}{the user input character for y-axis value}
}
\value{
a named list with variable names for visualization for internal use
}
\description{
Internal helper function that converts axis name to internal variable name
}
