% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{generate_pairdistance_balance_graph}
\alias{generate_pairdistance_balance_graph}
\title{Plotting function that generate sum of pair-wise distance v.s. total variation imbalance on balance variable}
\usage{
generate_pairdistance_balance_graph(matchingResult)
}
\arguments{
\item{matchingResult}{an object returned by the main matching function multiObjMatch}
}
\description{
Plotting function that generate sum of pair-wise distance v.s. total variation imbalance on balance variable
}
