% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{generate_pairdistance_graph}
\alias{generate_pairdistance_graph}
\title{Plotting function that generate sum of pair-wise distance v.s. number of unmatched treated units}
\usage{
generate_pairdistance_graph(matchingResult)
}
\arguments{
\item{matchingResult}{a matching result object containing information for all matches}
}
\description{
Plotting function that generate sum of pair-wise distance v.s. number of unmatched treated units
}
