% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{getExactOn}
\alias{getExactOn}
\title{Get a factor of string representing the exact variables.}
\usage{
getExactOn(dat, exactList)
}
\arguments{
\item{dat}{dataframe containing all the variables in exactList}

\item{exactList}{factor of names of the variables on which we want exact or close matching.}
}
\value{
factor of concatenated variable values for the variables we want exact or close matching on.
}
\description{
Get a factor of string representing the exact variables.
}
