% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{getPropensityScore}
\alias{getPropensityScore}
\title{Fit propensity scores using the given covariates}
\usage{
getPropensityScore(df, covs)
}
\arguments{
\item{df}{dataframe that contains a column named "treat", the treatment vector, and columns of covariates specified.}

\item{covs}{factor of column names of covariates used for fitting a propensity score model.}
}
\value{
factor of fitted propensity scores.
}
\description{
Fit propensity scores using the given covariates
}
