% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{getUnmatched}
\alias{getUnmatched}
\title{A function that generate the percentage of unmatched units for each match.}
\usage{
getUnmatched(matchingResult)
}
\arguments{
\item{matchingResult}{matchingResult object that contains information for all matches}
}
\value{
dataframe that contain number of matched units as well as the percentage of matched units
}
\description{
A function that generate the percentage of unmatched units for each match.
}
