% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiObjMatch.R
\name{matched_data}
\alias{matched_data}
\title{A function that returns the dataframe that contains only matched pairs from the original data frame with specified match index}
\usage{
matched_data(matchingResult, match_num)
}
\arguments{
\item{matchingResult}{an object returned by the main matching function multiObjMatch}

\item{match_num}{Integer index of match that the user want to extract paired observations from}
}
\value{
dataframe that contains only matched pair data
}
\description{
A function that returns the dataframe that contains only matched pairs from the original data frame with specified match index
}
